honey-bounce: low-priority code quality and documentation fixes
- SpringChain: Add SETTLING_THRESHOLD = 0.0005 constant
- SpringChain: Rename build() to new() per SugarCraft convention
- Point/Vector: Implement JsonSerializable
- Point/Vector: Add __toString() for debugging
- Spring: Change fps(int) to fps(float) for NTSC frame rates
- ReducedMotionTest: Add @group parallel-unsafe annotation
- Examples: Improve autoload path portability with dirname(__DIR__)

+++ src/Spring.php
@@ -127,9 +127,9 @@ final class Spring
     }
 
     /** Frame-time (seconds) for a given frame rate. `Spring::fps(60)` → 1/60. */
-    public static function fps(int $n): float
+    public static function fps(float $n): float
     {
-        if ($n <= 0) {
+        if ($n <= 0.0) {
             throw new \InvalidArgumentException(Lang::t('spring.fps_positive', ['fps' => $n]));
         }
         return 1.0 / $n;
